Update the WpsClient and WpsService to be able to create WpsScriptPlugins adding wps process to OrbisGIS.
