+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/ui/dataui/FieldValueUI.java
@@ -21,7 +21,7 @@ package org.orbisgis.wpsclient.view.ui.dataui;
 
 import net.miginfocom.swing.MigLayout;
 import org.orbisgis.commons.progress.SwingWorkerPM;
-import org.orbisgis.wpsclient.WpsClient;
+import org.orbisgis.wpsclient.WpsClientImpl;
 import org.orbisgis.wpsclient.view.utils.ToolBoxIcon;
 import org.orbisgis.wpsservice.model.*;
 
@@ -65,10 +65,10 @@ public class FieldValueUI implements DataUI{
     private static final String LAYERUI_PROPERTY = "LAYERUI_PROPERTY";
 
     /** WpsClient using the generated UI. */
-    private WpsClient wpsClient;
+    private WpsClientImpl wpsClient;
 
     @Override
-    public void setWpsClient(WpsClient wpsClient){
+    public void setWpsClient(WpsClientImpl wpsClient){
         this.wpsClient = wpsClient;
     }
